delegating all backend calls to crossterm
everything minus the backed::size that uses /dev/tty (which we won't
have on a ssh remote terminal context)

+++ src/ssh/app.rs
@@ -206,7 +206,7 @@ impl Handler for AppServer {
                 height: height as u16,
             };
             terminal.resize(rect)?;
-            terminal.backend_mut().resize(width, height);
+            terminal.backend_mut().resize(width as u16, height as u16);
         }
 
         Ok(())
